Adding TutorialsSidebarContent component
