+++ src/components/tutorials-sidebar/index.tsx
@@ -11,6 +11,9 @@ import {
   TutorialSidebarProps,
 } from './types'
 import s from './tutorials-sidebar.module.css'
+import { useCurrentProduct } from 'contexts'
+import useCurrentPath from 'hooks/use-current-path'
+import { CollectionCategorySidebarSection } from 'views/collection-view/helpers'
 
 function TutorialsSidebar({
   backToLinkProps,
@@ -29,6 +32,54 @@ function TutorialsSidebar({
     >
       {children}
     </Sidebar>
+  )
+}
+
+function TutorialsSidebarContent({
+  sections,
+}: {
+  sections: CollectionCategorySidebarSection[]
+}) {
+  const currentProduct = useCurrentProduct()
+  const currentPath = useCurrentPath({ excludeHash: true, excludeSearch: true })
+
+  return (
+    <>
+      <SectionList
+        items={[
+          {
+            text: 'Overview',
+            href: `/${currentProduct.slug}/tutorials`,
+            isActive: currentPath === `/${currentProduct.slug}/tutorials`,
+          },
+        ]}
+      />
+      {sections?.map((section: CollectionCategorySidebarSection) => {
+        return (
+          <>
+            <HorizontalRule />
+            <SectionTitle text={section.title} />
+            <SectionList items={section.items} />
+          </>
+        )
+      })}
+      <HorizontalRule />
+      <SectionTitle text="Resources" />
+      <SectionList
+        items={[
+          {
+            text: 'All Tutorials',
+            href: 'https://learn.hashicorp.com',
+          },
+          {
+            text: 'Community Forum',
+            href: 'https://discuss.hashicorp.com',
+          },
+          { text: 'Support', href: 'https://support.hashicorp.com' },
+          { text: 'GitHub', href: 'http://github.com/hashicorp' },
+        ]}
+      />
+    </>
   )
 }
 
@@ -61,5 +112,11 @@ function HorizontalRule() {
   return <SidebarHorizontalRule />
 }
 
-export { HorizontalRule, ListItem, SectionList, SectionTitle }
+export {
+  HorizontalRule,
+  ListItem,
+  SectionList,
+  SectionTitle,
+  TutorialsSidebarContent,
+}
 export default TutorialsSidebar
