+++ src/views/product-tutorials-view/index.tsx
@@ -1,6 +1,4 @@
-import useCurrentPath from 'hooks/use-current-path'
 import { useCurrentProduct } from 'contexts'
-import { CollectionCategorySidebarSection } from 'views/collection-view/helpers'
 import SidebarSidecarLayout from 'layouts/sidebar-sidecar'
 import CoreDevDotLayout from 'layouts/core-dev-dot-layout'
 import {
@@ -8,9 +6,7 @@ import {
   generateTopLevelSidebarNavData,
 } from 'components/sidebar/helpers'
 import TutorialsSidebar, {
-  HorizontalRule,
-  SectionList,
-  SectionTitle,
+  TutorialsSidebarContent,
 } from 'components/tutorials-sidebar'
 import { ProductTutorialsSitemap } from './components'
 import { ProductTutorialsViewProps } from './server'
@@ -24,11 +20,9 @@ function ProductTutorialsView({
   data,
 }: ProductTutorialsViewProps): React.ReactElement {
   const currentProduct = useCurrentProduct()
-  const currentPath = useCurrentPath({ excludeHash: true, excludeSearch: true })
   const { inlineCollections, inlineTutorials, pageData, allCollections } = data
   const { showProductSitemap, blocks } = pageData
 
-  // TODO: refactor, very long
   const sidebarNavDataLevels = [
     generateTopLevelSidebarNavData(currentProduct.name),
     generateProductLandingSidebarNavData(currentProduct),
@@ -42,46 +36,9 @@ function ProductTutorialsView({
         href: `/${currentProduct.slug}`,
       },
       title: 'Tutorials',
-      children: [
-        <>
-          <SectionList
-            items={[
-              {
-                text: 'Overview',
-                href: `/${currentProduct.slug}/tutorials`,
-                isActive: currentPath === `/${currentProduct.slug}/tutorials`,
-              },
-            ]}
-          />
-          {layoutProps.sidebarSections.map(
-            (section: CollectionCategorySidebarSection) => {
-              return (
-                <>
-                  <HorizontalRule />
-                  <SectionTitle text={section.title} />
-                  <SectionList items={section.items} />
-                </>
-              )
-            }
-          )}
-          <HorizontalRule />
-          <SectionTitle text="Resources" />
-          <SectionList
-            items={[
-              {
-                text: 'All Tutorials',
-                href: 'https://learn.hashicorp.com',
-              },
-              {
-                text: 'Community Forum',
-                href: 'https://discuss.hashicorp.com',
-              },
-              { text: 'Support', href: 'https://support.hashicorp.com' },
-              { text: 'GitHub', href: 'http://github.com/hashicorp' },
-            ]}
-          />
-        </>,
-      ],
+      children: (
+        <TutorialsSidebarContent sections={layoutProps.sidebarSections} />
+      ),
     },
   ]
